requirements.txt
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
customtkinter>=5.2.0
